% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_annon.R
\name{combine_tax_df}
\alias{combine_tax_df}
\title{Combine tax list}
\usage{
combine_tax_df(tax_list)
}
\arguments{
\item{tax_list}{the list containing multiple tax dataframes. Each tax dataframe needs to contain 7 columns (Kingdom,Phylum,Class,Order,Family,Genus,Species)}
}
\value{
A dataframe where all the taxonomic information is merged with " | ". Example tax_list[[1]] = "Gadus Morhua" & tax_list[[1]] = "Arctogadus glacialis" the result will be "Gadus Morhua | Arctogadus glacialis"
}
\description{
This function combines different tax dataframes (each as a different element on a list) into a singular dataframe
}
\examples{
comb_tax <- combine_tax_df(tax_list)

}
