% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_annon.R
\name{edna_to_fasta}
\alias{edna_to_fasta}
\title{eDNA data to fasta}
\usage{
edna_to_fasta(qid, seq, output = "out.fasta")
}
\arguments{
\item{qid}{The vector containig the query ID}

\item{seq}{The vector containig the sequences}

\item{output}{the name of the output file}
}
\value{
A fasta file with extenxtion .fasta
}
\description{
This function creates fasta sequences from edna reads
This is very handy when you want to reblast some sequences
}
\examples{
# edna_to_fasta(edna$id,edna$sequence,"Plate_NANB.fasta")
}
