% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_annon.R
\name{edna_to_otu}
\alias{edna_to_otu}
\title{Converts edna dataframe into OTU dataframe}
\usage{
edna_to_otu(edna_df)
}
\arguments{
\item{edna_df}{The edna dataframe}
}
\value{
OTU dataframe
}
\description{
This function selects only the columns where the samples are using "taxid" ans "seq_length"
}
\examples{
otu <- coll_sim2(edna_to_otu(edna),tax_to_vector(comb_tax_compared))
colSums(edna_to_otu(edna))
}
