% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_annon.R
\name{incos_comb_tax_df}
\alias{incos_comb_tax_df}
\title{Inconsistencies in combining tax lists}
\usage{
incos_comb_tax_df(comb_tax)
}
\arguments{
\item{comb_tax}{The combined tax dataframe created from combine_tax_df()}
}
\value{
A dataframe with only the rows where the taxonomic information has multiple information. Example, "Gadus Morhua | Arctogadus glacialis"
}
\description{
This function creates a dataframe with only the taxa that are not agreeing between blasted and local database results. This function is embedded in tax_blast_compare() function.
}
\examples{
inco_comb_tax <- incos_comb_tax_df(comb_tax)

}
