% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_annon.R
\name{Packages asd}
\alias{Packages asd}
\alias{manually_annotate}
\title{Annotate blast results}
\usage{
manually_annotate(
  input,
  method = 1,
  skip = T,
  print_blast = T,
  query_vector = NA
)
}
\arguments{
\item{input}{the blast output dataframe from NCBI}

\item{method}{The annotation method. 1=Fully manual; 2=Semi-automatic; 3=Fully Automatic}
}
\value{
prints a csv file in your working directory named "manually_annotated_function_outcome.csv"

returns a dataframe containing query id, species names of query match and % match
}
\description{
This function reads blasts results and automatically or manually annotates the blasted sequences
to the best match resutl
}
\examples{
manually_annotate(blast_output,method = 3)
}
