% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_annon.R
\name{pr_df}
\alias{pr_df}
\title{Print as dataframe}
\usage{
pr_df(vector_var, wide = 40, new_row = F)
}
\arguments{
\item{vector_var}{The variable/vector that you want to print}

\item{wide}{The number of wide space you want from the first variable to the next}
}
\value{
The formated output in Console
}
\description{
This function prints a multiple vectors in the form of a dataframe. Mainly for esthetic reasons
}
\examples{
pr_df(tax$id,50)
pr_df(tax$sequences,50,new_row = T)
sliced_sequences <- pr_df(tax$sequence,50) #if you want to slice all the variables in that vector to 50 characters 
cat(pr_df(tax$genus_name,30),pr_df(tax$species_name,30),pr_df(tax$taxid,10),sep = "\n")
}
