% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_annon.R
\name{rowpaste}
\alias{rowpaste}
\title{Row Paste}
\usage{
rowpaste(inn)
}
\arguments{
\item{inn}{the dataframe where you want the columns should be collapsed}
}
\value{
A vector of pasted materials
}
\description{
This function pastes rows one by one
}
\examples{
rowpaste(tax[,1:6])
metazoans <- rowpaste(tax[,tax$superkindom=="Metazoa"])
}
