% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_annon.R
\name{tax_blast_compare}
\alias{tax_blast_compare}
\title{Selecting the right tax info between combined results}
\usage{
tax_blast_compare(comb_tax)
}
\arguments{
\item{comb_tax}{The combined tax dataframe created from combine_tax_df()}
}
\value{
A dataframe where all the taxonomic information is merged with " | ". Example tax_list[[1]] = "Gadus Morhua" & tax_list[[1]] = "Arctogadus glacialis" the result will be "Gadus Morhua | Arctogadus glacialis"
}
\description{
This function is an interactive script asking for input on chosing the correct taxonomic information when the blasted and the local results are not agreeing
}
\examples{
comb_tax_compared <- tax_blast_compare(comb_tax)

}
