% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_annon.R
\name{tax_to_vector}
\alias{tax_to_vector}
\title{Creates a single vector with all the taxonomic info}
\usage{
tax_to_vector(comb_tax)
}
\arguments{
\item{comb_tax}{The combined tax dataframe created from combine_tax_df()}
}
\value{
A vector of length = nrow(input_df) where all the taxonomic information is concatenated into 1 string separated with "/"
}
\description{
This function collapses all the taxonomic information (all 7 columns mentioned in combine_tax_df()) into 1 string separated with "/". Can be used on dataframes as well where each row becomes one string of a vector (of length = nrow(df))
}
\examples{
comb_tax[!duplicated(tax_to_vector(comb_tax[,1:7])),]

}
